+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7PR6M8P03F8NCABBV7J6W0M"
+        "01K7RBQGRJZ0J5MXKNCAQ8SEF5"
     )
     r"""The id of the resource"""