+++ packages/orq-rc/RELEASES.md
@@ -3389,3 +3389,13 @@ Based on:
 - [python v3.14.0-rc.34] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.34] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.34 - packages/orq-rc
+
+## 2025-10-17 05:55:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.0 (2.727.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.35] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.35] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.35 - packages/orq-rc